casting to int

+++ datagrump-receiver.cc
@@ -28,13 +28,13 @@ int main( int argc, char *argv[] )
 			argv[ 1 ] /* port */ ) );
 
     fprintf( stderr, "Listening on port %s...\n", argv[ 1 ] );
-    fprintf( stderr, "Start On %d \n",timestamp());
+    fprintf( stderr, "Start On %d \n",(int)timestamp());
 
     /* Loop */
     uint64_t sequence_number = 0;
     while ( 1 ) {
       Packet received_packet = sock.recv();
-      fprintf( stderr, "Recv on %d\n",timestamp());
+      fprintf( stderr, "Recv on %d\n",(int)timestamp());
 
       /* Send back acknowledgment */
       Packet ack( received_packet.addr(), sequence_number++, received_packet );
